+++ packages/flutter/lib/src/integrations/integrations.dart
@@ -3,7 +3,7 @@ export 'flutter_error_integration.dart';
 export 'load_contexts_integration.dart';
 export 'load_debug_images_integration.dart';
 export 'load_release_integration.dart';
-export 'native_app_start_integration.dart';
+export '../app_start/ui_load_attached/native_app_start_integration.dart';
 export 'on_error_integration.dart';
 export 'sdk_integration.dart';
 export 'widgets_binding_integration.dart';